Moved alphabet into dataset
