--- pgse/etc/__init__.py
@@ -1,27 +1,0 @@
-from pgse.etc.alphabet import (
-    AUTO,
-    DNA,
-    DNA_CHARS,
-    DNA_COMPLEMENT,
-    Alphabet,
-    canonicalize,
-    get_alphabet,
-    get_complement,
-    is_canonical,
-    reset_alphabet,
-    set_alphabet,
-)
-
-__all__ = [
-    "AUTO",
-    "DNA",
-    "DNA_CHARS",
-    "DNA_COMPLEMENT",
-    "Alphabet",
-    "canonicalize",
-    "get_alphabet",
-    "get_complement",
-    "is_canonical",
-    "reset_alphabet",
-    "set_alphabet",
-]
