+++ pgse/genome/utils.py
@@ -1,3 +1,3 @@
-from pgse.etc.alphabet import canonicalize, get_complement, is_canonical
+from pgse.dataset.alphabet import canonicalize, get_complement, is_canonical
 
 __all__ = ["canonicalize", "get_complement", "is_canonical"]
