+++ pgse/pipeline/regular_pipline.py
@@ -4,7 +4,7 @@ import pandas as pd
 import ray
 
 from pgse.environment.ray_env import RayEnvManager
-from pgse.etc.alphabet import AUTO, Alphabet, AlphabetArg, ComplementArg, set_alphabet
+from pgse.dataset.alphabet import AUTO, Alphabet, AlphabetArg, ComplementArg, set_alphabet
 from pgse.log import logger
 from pgse.model.model_trainer import ModelTrainer
 from pgse.dataset.file_label import FileLabel
