+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "pgse"
-version = "0.10.4"
+version = "0.10.5"
 description = "Progressive Genome Segment Enhancement (PGSE)"
 readme = "README.md"
 requires-python = ">=3.10"
